create abstract NFtModel and seprate TezosNFTModel
