+++ lib/dashboard/home/tab_bar/nft/models/models.dart
@@ -1,1 +1,2 @@
 export 'nft_model.dart';
+export 'tezos_nft_model.dart';
